FAC-977: Rename GetAssessmentInsight to GetAssessmentOverallInsight
