+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/assessment/GetAssessmentOverallInsightRestController.java
@@ -2,10 +2,10 @@ package org.flickit.assessment.core.adapter.in.rest.insight.assessment;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.adapter.in.rest.insight.assessment.GetAssessmentInsightResponseDto.InsightDetail;
+import org.flickit.assessment.core.adapter.in.rest.insight.assessment.GetAssessmentOverallInsightResponseDto.InsightDetail;
 import org.flickit.assessment.core.application.domain.insight.Insight;
-import org.flickit.assessment.core.application.port.in.insight.assessment.GetAssessmentInsightUseCase;
-import org.flickit.assessment.core.application.port.in.insight.assessment.GetAssessmentInsightUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.assessment.GetAssessmentOverallInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.assessment.GetAssessmentOverallInsightUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,13 +16,13 @@ import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
-public class GetAssessmentInsightRestController {
+public class GetAssessmentOverallInsightRestController {
 
-    private final GetAssessmentInsightUseCase useCase;
+    private final GetAssessmentOverallInsightUseCase useCase;
     private final UserContext userContext;
 
     @GetMapping("assessments/{assessmentId}/overall-insight")
-    ResponseEntity<GetAssessmentInsightResponseDto> getAssessmentInsight(@PathVariable("assessmentId") UUID assessmentId) {
+    ResponseEntity<GetAssessmentOverallInsightResponseDto> getAssessmentInsight(@PathVariable("assessmentId") UUID assessmentId) {
         var currentUserId = userContext.getUser().id();
         var result = toResponse(useCase.getAssessmentInsight(toParam(assessmentId, currentUserId)));
         return new ResponseEntity<>(result, HttpStatus.OK);
@@ -32,8 +32,8 @@ public class GetAssessmentInsightRestController {
         return new Param(assessmentId, currentUserId);
     }
 
-    private GetAssessmentInsightResponseDto toResponse(Insight insight) {
-        return new GetAssessmentInsightResponseDto(toInsightDetail(insight.defaultInsight()),
+    private GetAssessmentOverallInsightResponseDto toResponse(Insight insight) {
+        return new GetAssessmentOverallInsightResponseDto(toInsightDetail(insight.defaultInsight()),
             toInsightDetail(insight.defaultInsight()),
             insight.editable(),
             insight.approved());
